cleanup json parser
