+++ src/libjson/sample2/Makefile
@@ -3,8 +3,10 @@ CFLAGS = -Wall
 INCLUDES = -I../include
 LDFLAGS = -L.. -ljson
 
+SOURCES = main.c ../common/load_file.c
+
 $(TARGET): main.c
-	gcc $(CFLAGS) $(INCLUDES) -o $(TARGET) main.c $(LDFLAGS)
+	gcc $(CFLAGS) $(INCLUDES) -o $(TARGET) $(SOURCES) $(LDFLAGS)
 
 clean:
 	rm -f $(TARGET)
