+++ src/libjson/sample2/main.c
@@ -33,6 +33,7 @@
 #include <json.h>
 #include <sys/stat.h>
 #include <string.h>
+#include "../common/load_file.h"
 
 const char* arg0;
 
@@ -226,60 +227,6 @@ static json_result_t _callback(
     return JSON_OK;
 }
 
-static int _load_file(
-    const char* path,
-    size_t extra_bytes,
-    void** data_out,
-    size_t* size_out)
-{
-    int ret = -1;
-    FILE* is = NULL;
-    void* data = NULL;
-    size_t size;
-
-    /* Get size of this file */
-    {
-        struct stat st;
-
-        if (stat(path, &st) != 0)
-            goto done;
-
-        size = (size_t)st.st_size;
-    }
-
-    /* Allocate memory */
-    if (!(data = malloc(size + extra_bytes)))
-        goto done;
-
-    /* Open the file */
-    if (!(is = fopen(path, "rb")))
-        goto done;
-
-    /* Read file into memory */
-    if (fread(data, 1, size, is) != size)
-        goto done;
-
-    /* Zero-fill any extra bytes */
-    if (extra_bytes)
-        memset((unsigned char*)data + size, 0, extra_bytes);
-
-    *data_out = data;
-    *size_out = size;
-    data = NULL;
-
-    ret = 0;
-
-done:
-
-    if (data)
-        free(data);
-
-    if (is)
-        fclose(is);
-
-    return ret;
-}
-
 static void _trace(
     json_parser_t* parser,
     const char* file,
@@ -308,7 +255,7 @@ static void _parse(const char* path)
     cd.newline = 0;
     cd.comma = 0;
 
-    if (_load_file(path, 1, (void**)&data, &size) != 0)
+    if (json_load_file(path, 1, (void**)&data, &size) != 0)
     {
         fprintf(stderr, "%s: failed to access '%s'\n", arg0, path);
         exit(1);
